place order now do all necessary tasks
- send email, dispatch message, edit customer delivery address. subscribe to newsletter, ...
