+++ packages/framework/src/Model/Customer/User/CustomerUserFacade.php
@@ -13,7 +13,9 @@ use Shopsys\FrameworkBundle\Model\Customer\Customer;
 use Shopsys\FrameworkBundle\Model\Customer\CustomerDataFactoryInterface;
 use Shopsys\FrameworkBundle\Model\Customer\CustomerFacade;
 use Shopsys\FrameworkBundle\Model\Customer\DeliveryAddress;
+use Shopsys\FrameworkBundle\Model\Customer\DeliveryAddressDataFactory;
 use Shopsys\FrameworkBundle\Model\Customer\DeliveryAddressFacade;
+use Shopsys\FrameworkBundle\Model\Customer\DeliveryAddressFactory;
 use Shopsys\FrameworkBundle\Model\Customer\Exception\DuplicateEmailException;
 use Shopsys\FrameworkBundle\Model\Customer\Mail\CustomerMailFacade;
 use Shopsys\FrameworkBundle\Model\Order\Order;
@@ -33,6 +35,8 @@ class CustomerUserFacade
      * @param \Shopsys\FrameworkBundle\Model\Customer\CustomerDataFactoryInterface $customerDataFactory
      * @param \Shopsys\FrameworkBundle\Model\Customer\BillingAddressFacade $billingAddressFacade
      * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUserRefreshTokenChainFacade $customerUserRefreshTokenChainFacade
+     * @param \Shopsys\FrameworkBundle\Model\Customer\DeliveryAddressFactory $deliveryAddressFactory
+     * @param \Shopsys\FrameworkBundle\Model\Customer\DeliveryAddressDataFactory $deliveryAddressDataFactory
      */
     public function __construct(
         protected readonly EntityManagerInterface $em,
@@ -47,6 +51,8 @@ class CustomerUserFacade
         protected readonly CustomerDataFactoryInterface $customerDataFactory,
         protected readonly BillingAddressFacade $billingAddressFacade,
         protected readonly CustomerUserRefreshTokenChainFacade $customerUserRefreshTokenChainFacade,
+        protected readonly DeliveryAddressFactory $deliveryAddressFactory,
+        protected readonly DeliveryAddressDataFactory $deliveryAddressDataFactory,
     ) {
     }
 
@@ -319,5 +325,74 @@ class CustomerUserFacade
     public function getByUuid(string $uuid): CustomerUser
     {
         return $this->customerUserRepository->getOneByUuid($uuid);
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
+     * @param \Shopsys\FrameworkBundle\Model\Order\Order $order
+     * @param string|null $deliveryAddressUuid
+     * @param bool $isSubscribeToNewsletter
+     */
+    public function updateCustomerUserByOrder(
+        CustomerUser $customerUser,
+        Order $order,
+        ?string $deliveryAddressUuid,
+        bool $isSubscribeToNewsletter,
+    ): void {
+        $deliveryAddress = $this->resolveDeliveryAddress($deliveryAddressUuid, $customerUser);
+        $customerUserUpdateData = $this->customerUserUpdateDataFactory->createFromCustomerUser($customerUser);
+        $customerUserUpdateData->customerUserData->newsletterSubscription = $isSubscribeToNewsletter;
+        $this->editByCustomerUser($customerUser->getId(), $customerUserUpdateData);
+        $deliveryAddress = $deliveryAddress ?? $this->createDeliveryAddressForAmendingCustomerUserData($order);
+        $this->amendCustomerUserDataFromOrder($customerUser, $order, $deliveryAddress);
+    }
+
+    /**
+     * @param string|null $deliveryAddressUuid
+     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
+     * @return \Shopsys\FrameworkBundle\Model\Customer\DeliveryAddress|null
+     */
+    protected function resolveDeliveryAddress(
+        ?string $deliveryAddressUuid,
+        CustomerUser $customerUser,
+    ): ?DeliveryAddress {
+        if ($deliveryAddressUuid === null) {
+            return null;
+        }
+
+        return $this->deliveryAddressFacade->findByUuidAndCustomer(
+            $deliveryAddressUuid,
+            $customerUser->getCustomer(),
+        );
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Order\Order $order
+     * @return \Shopsys\FrameworkBundle\Model\Customer\DeliveryAddress|null
+     */
+    protected function createDeliveryAddressForAmendingCustomerUserData(Order $order): ?DeliveryAddress
+    {
+        $orderTransport = $order->getTransportItem()->getTransport();
+
+        if (
+            $orderTransport->isPersonalPickup() ||
+            $orderTransport->isPacketery() ||
+            $order->isDeliveryAddressSameAsBillingAddress()
+        ) {
+            return null;
+        }
+
+        $deliveryAddressData = $this->deliveryAddressDataFactory->create();
+        $deliveryAddressData->firstName = $order->getDeliveryFirstName();
+        $deliveryAddressData->lastName = $order->getDeliveryLastName();
+        $deliveryAddressData->companyName = $order->getDeliveryCompanyName();
+        $deliveryAddressData->street = $order->getDeliveryStreet();
+        $deliveryAddressData->city = $order->getDeliveryCity();
+        $deliveryAddressData->postcode = $order->getDeliveryPostcode();
+        $deliveryAddressData->country = $order->getDeliveryCountry();
+        $deliveryAddressData->postcode = $order->getDeliveryPostcode();
+        $deliveryAddressData->customer = $order->getCustomerUser()?->getCustomer();
+
+        return $this->deliveryAddressFactory->create($deliveryAddressData);
     }
 }
